Add GetStateForActionHandlers
